move useCategoryModal hook to each hooks folder
