+++ client/app/(main)/category/components/CategoryDisplay.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
 import { Categories, CategoryDeleteModal, CategoryUpdateModal } from '@/app/(main)/category/components'
+import { useCategoryModal } from '@/app/(main)/category/hooks'
 import { Category } from '@/app/types'
-import { useCategoryModal } from '@/app/utils'
 
 const CategoryDisplayContainer = styled.div`
   overflow-y: scroll;
