+++ client/app/(main)/category/hooks/index.ts
@@ -1,0 +1,3 @@
+'use client'
+
+export * from './useCategoryModal'
