+++ client/app/utils/hooks/index.ts
@@ -1,6 +1,5 @@
 'use client'
 
-export * from './category/useCategoryModal'
 export * from './drag/useDragFirst'
 export * from './drag/useDragDndKit'
 export * from './todolist/useTodolist'
